Changed names for method parts to match the Simple Theme, now we have:
 1. declaration.function -- the entire thing
 2.  function-result     -- result of function
 3.  function-name       -- name of funtcion
 4.  function-arguments  -- the argument variable
 5.  function-arg-type   -- the type of the argument value
 6.  named-argument      -- the naem of the parameter

All languages should support 1, 3 and 4.
Manifest typed langauges will support 2 and 5.
Cool languages will support 6.

Changes to the names are welcome. I think it should “function-argument” instead of the plural form.

Also, “function-arg-type” looks a bit out-of-place. The reason for going with arg here is the prefix matching of selectors.



git-svn-id: http://svn.textmate.org/trunk/Bundles/Objective-C.tmbundle@643 dfb7d73b-c2ec-0310-8fea-fb051d288c6d