Adding a new “Common Method Calls” snippet to register for an NSNotification.
The notification name is transformed into the method name by removing an optional 2 uppercase-character prefix and optional “Notification” suffix, so e.g. NSWindowDidBecomeMainNotification → windowDidBecomeMain:
Alternatively I could have tabstopped the method name first and transformed this into the notification name, but I think the most common scenario is going to be copying the notification name from the docs or some header and pasting it in (and also the method name is customisable whereas the notification name is not).


git-svn-id: http://svn.textmate.org/trunk/Bundles/Objective-C.tmbundle@9550 dfb7d73b-c2ec-0310-8fea-fb051d288c6d