Extend documentation lookup to C++ data types/functions.
This is done by referencing a local copy of www.cppreference.com.

The reason I included a local copy (as opposed to referencing the online pages) is that mapping a query to a file is done by searching the file system. Ideally though we’ll replace the local mirror with an index.