Give man page query results a proper title.
Previously we always used ‘man(2, 3)’ as the title for a man page query. Now we use ‘read(2)’ if the query was for ‘read’ (and the word was found in section 2).

The code can easily be changed to show all man pages for the word (by adding ‘-a’ to man) but I did not do that since we only query section 2 or 3, which should not have the same words, and by querying 2:3 for e.g. ‘open’ we get ‘open(3pm)’ as a result.