Order completion suggestions alphabetically
Results are shown in the order they are found in the input file(s). This commit ensures that these files are sorted.

It breaks if results are found in multiple files. Ideally the script will sort the results found, though it calls out to zgrep in a lot of places (and by different scripts), so I’d prefer to cleanup the code a bit first (there seems to be a lot of unused code which can be dropped, for example a Dialog v1.x version of the completion command).