Always place functions in the grammar repository
Currently the C and Objective-C grammars expect the functions to be in the repository for the C platform functions but not the Objective-C platform functions, this is a little inconsistent, so since having the functions in the repository is the most flexible, this is what we choose.

This commit also include the required change to the Objective-C grammar.