Fix sort/uniq of completion results
The current code was sorting on the filter string, but for methods with multiple arguments, the filter string does not contain the parameter names even though they are part of the display string.

The code to prune duplicates may have worked, but it assumed a sorted list, which was not always the case.

I moved sorting to the show_dialog function and made a (syntactically) simpler uniq function.